Cart download tmp file check

+++ app/jobs/cart_download_job.rb
@@ -55,19 +55,21 @@ class CartDownloadJob
       end
     end
 
-    if self.large
-      time = Time.now.to_i
-      large_path = "#{Rails.application.config.tmp_path}/large/#{sess_id}"
-      full_large_path = "#{large_path}/#{time}.zip"
+    if File.exists?(temp_path) #if not, most likely deleted in tmp sweep, and this is a stale request anyway
+      if self.large
+        time = Time.now.to_i
+        large_path = "#{Rails.application.config.tmp_path}/large/#{sess_id}"
+        full_large_path = "#{large_path}/#{time}.zip"
 
-      FileUtils.mkdir_p large_path
-      FileUtils.mv(temp_path, full_large_path)
+        FileUtils.mkdir_p large_path
+        FileUtils.mv(temp_path, full_large_path)
 
-      # Email user their download link
-      LargeDownloadMailer.download_alert(time, self.nuid, self.sess_id).deliver!
-    else
-      # Rename temp path to full path so download can pick it up
-      FileUtils.mv(temp_path, full_path)
+        # Email user their download link
+        LargeDownloadMailer.download_alert(time, self.nuid, self.sess_id).deliver!
+      else
+        # Rename temp path to full path so download can pick it up
+        FileUtils.mv(temp_path, full_path)
+      end
     end
   end
 end
